feat(plasma-core): add price periodicity
