+++ packages/plasma-core/src/components/Price/Price.tsx
@@ -1,31 +1,9 @@
 import React from 'react';
 import styled, { css } from 'styled-components';
 
-import type { AsProps } from '../../types';
 import { formatCurrency } from '../../utils';
 
-export interface PriceProps extends React.HTMLAttributes<HTMLSpanElement>, AsProps {
-    /**
-     * Значение контрола.
-     */
-    children: number;
-    /**
-     * Валюта.
-     */
-    currency?: 'rub' | 'usd' | 'eur';
-    /**
-     * Перечеркнутый текст (старая цена, например).
-     */
-    stroke?: boolean;
-    /**
-     * Минимальное количество цифр после десятичного разделителя.
-     */
-    minimumFractionDigits?: number;
-}
-
-interface StyledPriceProps {
-    $stroke?: boolean;
-}
+import { PriceProps, StyledPriceProps } from './types';
 
 const StyledPrice = styled.span<StyledPriceProps>`
     ${({ $stroke }) =>
@@ -33,6 +11,10 @@ const StyledPrice = styled.span<StyledPriceProps>`
         css`
             text-decoration: line-through;
         `};
+`;
+
+const StyledPeriodicity = styled.span`
+    padding-left: 0.25rem;
 `;
 
 /**
@@ -43,9 +25,11 @@ export const Price: React.FC<PriceProps> = ({
     currency = 'rub',
     minimumFractionDigits = 0,
     stroke = false,
+    periodicity,
     ...rest
 }) => (
     <StyledPrice $stroke={stroke} {...rest}>
         {formatCurrency(children, currency, minimumFractionDigits)}
+        {periodicity && <StyledPeriodicity>{periodicity}</StyledPeriodicity>}
     </StyledPrice>
 );
