+++ packages/plasma-core/src/components/Price/index.ts
@@ -1,2 +1,2 @@
 export { Price } from './Price';
-export type { PriceProps } from './Price';
+export type { PriceProps } from './types';
